fix(memory): user-reported bugs in Track-M/O
- bank_persistence: persist num_kv_heads_per_layer in save_bank meta,
  load_bank state_dict reconstruction, and compute_config_sha so
  heterogeneous-KV-head models (Gemma-4 [16,4,...]) survive round-trip
  and have a unique config_sha.
- bank_decay._filter_rows: blocklist per-layer metadata keys (head_dims,
  num_kv_heads_per_layer) so they are never truncated as if they were
  per-row when n_facts == num_layers.
- bank_tiering: add dtype=torch.long to demote('hot', i) keep tensor;
  empty list previously yielded float tensor, crashing index_select.
- pyproject.toml: add [obs] optional extra (prometheus-client, otel-api,
  otel-sdk) so docs/observability.md install command works.
- tests: add 5 regression tests covering each bug.

+++ deltamemory/memory/bank_tiering.py
@@ -46,7 +46,7 @@ class BankTier:
             if tier == "hot":
                 row_k = self.hot_k[i : i + 1].detach().cpu()
                 row_v = self.hot_v[i : i + 1].detach().cpu()
-                keep = torch.tensor([j for j in range(self.hot_k.size(0)) if j != i], device=self.hot_k.device)
+                keep = torch.tensor([j for j in range(self.hot_k.size(0)) if j != i], dtype=torch.long, device=self.hot_k.device)
                 self.hot_k = self.hot_k.index_select(0, keep)
                 self.hot_v = self.hot_v.index_select(0, keep)
                 self.warm_k = torch.cat([self.warm_k, row_k], dim=0)
